discovery/helios: use bootstrap discovery endpoints instead of hardcoded endpoint
